Implement ServerConfig
- Implement DatabaseManager
- Implement UserManager
- Implement Packet_Udp
- Implement UDPServer
- Implement ServerTick

+++ managers/packet_hostmanager.h
@@ -4,8 +4,8 @@
 class Packet_HostManager {
 public:
 	void ParsePacket_Host(TCPConnection::Packet::pointer packet);
-	void SendPacket_Host_StartGame(TCPConnection::pointer connection, int userID);
-	void SendPacket_Host_JoinGame(TCPConnection::pointer connection, int userID);
+	void SendPacket_Host_StartGame(TCPConnection::pointer connection, unsigned long userID);
+	void SendPacket_Host_JoinGame(TCPConnection::pointer connection, unsigned long userID);
 };
 
 extern Packet_HostManager packet_HostManager;
